Fix 1295677: REGRESSION: Import SETS startup page to imported home page.
We had only kHomePage pref before, which if blank/set to chrome-internal: meant that the New Tab page should be shown when Home is pressed. Now we add a pref called kHomePageIsNewTabPage for this purpose, which allows us to set the kHomePage value during import without actually using it as the HomePage (effectively importing home page becomes importing a default suggestion for when the users want a home page instead of the New Tab page).

BUG=1295677


git-svn-id: http://src.chromium.org/svn/trunk/src@107 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3d25decbf0cb00adc0cfaf17c2277148eaa11e11