Eliminate the infinite read loop that can occasionally allow a large, fast network request to monopolize the IO thread.
This particular bug was tickled by an FTP download where
pause tasks posted to the IO thread from the UI thread were
blocked from running until the FTP transfer was completed,
leading to user noticeable jankiness.

I added histograms to the previous read loop and noticed
that during regular browsing and downloading sessions, the
number of times the loop was traversed was 1 for something
like 99.5% of the time. This leads me to believe that
eliminating the read loop in favor of asynchronous tasks
will not impact performance.

BUG=1270179

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@212 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9957a202a16982197ba2c9ea1d3f1b527475f5d2