Revise the ObjectWatcher API to be one-to-one with the object being watched.  This greatly simplifies the implementation and API.
Now consumers can use ObjectWatcher as a "smart pointer" class, which automatically cleans-up after itself when it goes out of scope.

I also switched away from the Task based API to one that is more compatible with MessageLoop::WatchObject.  That allows me to make minimal changes to existing code, and it also means that consumers do not have to allocate Task objects to use this API.

In this CL, I included changes to make a couple consumers use ObjectWatcher instead of ML::WatchObject.

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@322 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3af2ca39222a4dc0dc2442e1cc1c8725d82fdfcc