Bring up the new frame (opaque version for XP only, for now).
I've hidden this frame behind a command line switch (--magic_browzR) so as not to destabilize the main browser UI any further. Note that running with this switch is likely buggy, incomplete, crashy, etc.

In order to make this work without disrupting a lot of existing code, I've had to make another BrowserView class (temporary) - BrowserView2. This also has to be a BrowserWindow implementor since that's the interface Browser uses to communicate with the UI.

BrowserView2 and OpaqueNonClientView are the major new files in this CL, but BrowserView2 is pretty similar to BrowserView. OpaqueNonClientView is the view that renders the titlebar/borders/etc. It's layout/painting routines are a bit simpler than XPFrame's!

B=1031854

git-svn-id: http://src.chromium.org/svn/trunk/src@329 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7614d8f0379d5495665918ecf1349aa19fc56dfd