Revised r408 such that we no longer sometimes fail MetricsServiceTest.CrashRenderers.
That test showed that we were exiting the browser process without noticing that a child process had crashed.  I fixed that by simply adding a check to see if the child process is still around before initiating the normal close of the child process.  This corresponds to the case where the browser decides to close a renderer.

As a result, the race for the MessageLoop to receive a Task notifying the RenderProcessHost of a crashed renderer is removed from the picture.

The bulk of this CL is just re-landing r408, which jar reviewed.

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@498 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ed33473900644929aa5362ed488b1e53a6dc4244