Save memory prefs to disks more often
The goal is to reduce the window where it is possible that a crash of Chrome
will cause double counting of prior crashes.
This change records all the (hopefully consistent) memory-resident state of the
metrics service to disk as soon as we get an ack from UMA that a
log was received.
Without this, there is a possible gap between when the stats are copied
to a log, and sent, vs. when the (now) zeroed crash count is written to disk.


bug=1314946

r=evanm,petersont

git-svn-id: http://src.chromium.org/svn/trunk/src@543 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5d52e9d5de0d6810b0ef34acbd2fdefc4f7b8cb9