Changed the Chrome Visual Studio solution to use the new Visual Studio project files for V8.
The following projects have been removed from the Chrome solution:

  libv8core
  no_snapshotv8
  mksnapshot
  snapshotv8
  v8_shell

The following projects (from v8/tools/visual_studio) have been added to the
Chrome solution:

  v8_base
  v8
  v8_mksnapshot
  v8_snapshot

together with the following dependencies:

  v8 depends on v8_base
  v8_mksnapshot depends on v8
  v8_snapshot on v8_base and v8_mksnapshot
  chrome_dll depends on v8_snapshot
  test_shell depends on v8_snapshot
  test_shell_tests depends on v8_snapshot
  perf_tests depends on v8_snapshot
  unit_tests depends on v8_snapshot
  generate_profile depends on v8_snapshot

Got rid of the IncrediBuild configuration file chrome.ib_ini as it is not needed with the new Visual Studio project files for V8.

The v8_shell project is not added to the Chrome solution as it will not be part of the V8 repository on code.google.com.


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@644 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: dbf7c00b5e6716f7e212ac8680873f2da19cb48e