Allow the ChromeViews' Menu to wrap the Windows system menu and insert items into it.
Basically adds an alternate ctor for Menu that takes an existing HMENU, and then provides variants of most of the APIs that allow insertion of an item at a specified index.

I will use this in the new frames to add the task manager and app frame menu items.

B=1293995

git-svn-id: http://src.chromium.org/svn/trunk/src@661 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c4baa48e01f9a8d8b2639d9b590cbd044116433e