Cleanup spell check file to stop using reference counted pointers without need.
I don't really believe this will solve anything (in terms
of the bug), but it did remove some needless complexity
above the crash site in the stack.  There is a tiny chance
that there is a shutdown race, where the destruction of the
containing object raced ahead of the execution (and
resulting destruction) of a task.  I think the destruction
of the task by the containing object would probably be
bad anyway, and I doubt that it even happens... so this
is all about being cleaner. 
bug=1318039

r=brettw

git-svn-id: http://src.chromium.org/svn/trunk/src@850 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b8996b8efcdb0907733174a80b2f0acb25d65edd