Temporaly added local copy of mjsunit.js to debugger tests.
The Chrome debugger tests use mjsunit.js from the V8 test framework. In the transition to a new structure for V8 on code.google.com the location of this file will change. A local copy is added temporaly to the debugger tests directory which can be removed when Chrome starts to use V8 from code.google.com.

git-svn-id: http://src.chromium.org/svn/trunk/src@867 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fe29a1c623ae412e57c798df2764d77413f4f306