This change fixes http://b/1325836, which is obscure but has a benefit of letting us pass another layout test: LayoutTests/editing/pasteboard/copy-standalone-image.html
It would probably be good to send this upstream eventually, because WebKit/Win doesn't pass this test, either.

WebKit/Mac for some reason inserts an extra P element around the image (along with margin-zero-setting style attribute to boot), but neither FF nor IE do that, so I decided to stick with the majority.

The test was rebaselined to a) remove the extra P element in the node tree and b) take minor pixel shifts due to Chrome rendering changes into the account.


git-svn-id: http://src.chromium.org/svn/trunk/src@939 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e921de96cd93443f902c779ac954f62479f8ec55