Copy from http://chrome-reviews.prom.corp.google.com/1237 (new gcl changelist model).
Description was:
Conditionally include personalization/ code by surrounding
the hooks into this module with #ifdef CHROME_PERSONALIZATION
in various .h/.cc files.

Building with the module requires adding this macro as a preprocessor
definition in build/internal/essential.vsprops, and adding it to the
VCResourceCompiler tool's command line (using /d). We will try and
make this easier in the future.

TBR=darin


git-svn-id: http://src.chromium.org/svn/trunk/src@955 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 899562d9c07a0f16fa09088940089f4970905745