Code review changes. Incorporated all the suggestions from previous review.
One item not changed yet is to rename 'receiver' in ForwardMessageToExternalHost.
The idea is to invoke receiver("message") at the other end. So for example if the
args to ForwardMessageToExternalHost("hello", "world") then we will invoke a
script hello("world") on the other side.

'receiver' doesn't really describe the first argument here so if there is a 
better suggestion, I would be happy to change it :)


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@1176 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4729ceb05d87431911e4fa4e4c35409303ccdbef