Fixed bug #1251296
This is a regression that is causing me to wonder why it ever worked in
the first place. The underflow bug was getting tripped by a very small scaling
matrix being improperly treated as a zero matrix, and also, the scaling code
was getting bitten by a bug in Skia's edge comparison function which caused
incorrect results to be returned when the difference between two values
exceeded the maximum signed integer.


git-svn-id: http://src.chromium.org/svn/trunk/src@1180 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e3c35dbc79b81a842f0a4fc5b8b021da1a45d4ec