Makes sure that debug-only layout test failures are not to the ZERO WIDTH SPACE mapping to SPACE glyph complaints (http://b/1317563), fixes a layout test (fast/text/zero-width-characters.html), and provides an updated patch for WebKit.org bug 20237 (https://bugs.webkit.org/show_bug.cgi?id=20237).
This change brings handling of the ZWS and CJK character widths down to the level of SimpleFontData by creating special (sub-classed) SimpleFontData objects that are used in GlyphData. These instances are created when the glyph cache is being filled (GlyphPage::fill). More better things are possible, but at the moment I thought it might be good to just get the basics right.

Also, a couple of the layout tests are brought back to pre-font-metric-hacks removal versions.


git-svn-id: http://src.chromium.org/svn/trunk/src@1557 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 206bd18c48b99c9d2fe1029716999cb1447fd88b