Fix proxy code. The pac string was corrupted when a pac file was specified.
This is because AsciiToWide returns a temporary object, and
we use .c_str, but as soon as the line is executed, the
object is destroyed and what was pointed by c_str is also
deleted.

Use a local variable instead.

Thanks for griffinz@gmail.com for pointing the issue.

Review URL: http://codereview.chromium.org/1838

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@1890 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 90a967300a628dcaea954f7a65e7bc5b884cd27d