Fix proxy code. The pac string was corrupted when a pac file was specified.
This is because AsciiToWide returns a temporary object, and
we use .c_str, but as soon as the line is executed, the
object is destroyed and what was pointed by c_str is also
deleted.

Use a local variable instead.

Thanks for griffinz@gmail.com for pointing the issue.

Review URL: http://codereview.chromium.org/1838

git-svn-id: http://src.chromium.org/svn/trunk/src@1890 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 733484a660825842cd64cdd938983fc9b03d8acc