This fixes http://code.google.com/p/chromium/issues/detail?id=643, which is an issue with the shockwave game not loading.
The plugin issues a number of geturlnotify requests. Some of these requests
come in with response headers without the content length. We would pass in -1
in the NPStream passed down to the plugin. As the end field in the NPStream
structure is an unsigned integer, the plugin would expect more data to arrive
and hence the issue.

The fix is to emulate Safari behavior and set the end field to 0 if the length
is -1.

Bug=643

Review URL: http://codereview.chromium.org/1676

git-svn-id: http://src.chromium.org/svn/trunk/src@1958 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 246c231f4ccc70fb6722983874831f771c190c93