Refactor the SSL tunnel establishment code so that it reuses the four states of writing the request headers and reading the response headers.
Add a boolean member is_connect_method_ to tell, while
we're in one of those four states, whether we're doing
the HTTP CONNECT method (to set up a tunnel) or the
real transaction.

The motivation of this refactoring is that I need to
read the response body when the CONNECT request fails,
which makes me realize I'd need to duplicate more code
with the original design.

R=darin
BUG=b/1272555
Review URL: http://codereview.chromium.org/1875

git-svn-id: http://src.chromium.org/svn/trunk/src@2011 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a55af926ad1605a7d9ed1aeda81b89cccf634ff5