Refactor the SSL tunnel establishment code so that it reuses the four states of writing the request headers and reading the response headers.
Add a boolean member is_connect_method_ to tell, while
we're in one of those four states, whether we're doing
the HTTP CONNECT method (to set up a tunnel) or the
real transaction.

The motivation of this refactoring is that I need to
read the response body when the CONNECT request fails,
which makes me realize I'd need to duplicate more code
with the original design.

R=darin
BUG=b/1272555
Review URL: http://codereview.chromium.org/1875

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2011 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5d2fd5c19956d92b3669c8e4f1340d23635cd1a4