Stop exposing manually_selected_match_ outside the AutocompletePopupModel.  The main goal of this is to be a first step towards divorcing the popup and the edit from each other.
To do this, I changed the behavior of manual selections.  They now do not persist once the user types more characters, hits esc, etc.  Our old behavior, which Brett and I designed long ago, turns out to have been a mistake; users who arrowed to an item and then typed more weren't expecting "stickiness" on their previous choice, and it led to user mistakes.

This also required changing how we do the "keyword UI" persistence in the case where the user switches into keyword UI, but then deletes all his text.  Previously, we used manually_selected_match_ with a provider affinity to the keyword provider in order to accomplish this.  Now we stick another flag on the AutocompleteInput, which, when set, biases the keyword provider to return the best results.  The user-visible effect of this is that when in keyword UI mode with no query string, the selected entry in the popup will be the first, rather than third, entry.  This is a small win.

While here I fixed the bug where editing a string and transforming it into a keyword search would avoid switching into keyword UI (as expected), but also delete the keyword off the visible string (oops).  I also made us lock the popup once the user changes the manually_selected_match_, in order to give a little more stability to it.

I'm sorry this makes so many behavioral changes at once.  All this code is tangled together and untangling it is hard :(

The keyword-related variables in the AutocompleteEditModel seem a mess.  They are probably worse now than before this change; I think I need a followup change at some point to make them all more sane.  It seems like we have three variables and complex conditionals where two, and simpler ones, would do.

BUG=997976,1201974,1204173
Review URL: http://codereview.chromium.org/3172

git-svn-id: http://src.chromium.org/svn/trunk/src@2426 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 120be768db6795e1c64873660fed5242139354a1