Correct typo in SDCH (filter chaining) unit test, and enable test
A line got deleted between testing and landing, which caused
the chaining test to break.  I disabled the chaining test
when the tree went red, but this CL has the one-line fix (to
the test).


I also found that the last series of edits (switching to
using status of FILTER_NEED_MORE_DATA rather than an empty
input buffer) was not completely tested :-(.  This also
explains how I added the typo to the test but did not
notice it.  :-(  :-(   This too had a required one line
of editing (in the filter constructor to use that status at
startup).

I then added a test line to catch the constructor problem,
and did a bit of lint cleanup.

Remaining TODO items center on:
a) Security restriction changes and/or tests.
b) Error recovery from network stripping of protocol headers
(some proxy servers may strip headers, and require SDCH to
be disabled along with output a meta-refresh page (hack)).

r=huanr
Review URL: http://codereview.chromium.org/4219

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2582 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 62cac633674af5ed7f83ae7a8d62c9fa1ea75b13