Adds the ability to sort TableView. Contrary to what we spoke about the other day I ended up doing the sorting in tableview. This makes it a heck of lot easier than having every model have to deal with it. As part of this I removed the optional non-caching logic from TableView, which was never used. Sadly though, this means there are coordinate transformations.
I've only enabled sorting in the keyword editor, I have to make sure
all the other places that use TableView can deal with it. For example,
task manager can't deal with it currently as it expects the getters to
be called only once where as when sorting they may be called multiple
times.

BUG=2790
TEST=This enables sorting ONLY in the keyword editor. Make sure there
  aren't any problems in adding/removing/changing entries in the
  keyword editor after this.

Review URL: http://codereview.chromium.org/4276

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2631 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2d6c6b21f378c0fe9f83cd93507399827368cf14