This fixes http://code.google.com/p/chromium/issues/detail?id=3007, which is the mouse-click-events.html layout test failure. Fix as per below:-
1. The test shell mouse event generating mechanism now tracks the last mouse button number received. This is used to determine whether the click count continues to increment or not. 

2. We now default to middle button if the button number passed in from js contains anything other than 0 or 2. This is as per the event sending mechanism in webkit mac.

We still need to re-baseline this test as the number of events printed on the page cause a scrollbar to show up for the 4th mouse event test. This causes the middle mouse button to be eaten as per recent changes in EventHandler.cpp. 

Bug=3007
R=tony

Review URL: http://codereview.chromium.org/6261

git-svn-id: http://src.chromium.org/svn/trunk/src@2917 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 23d7d02eaf6793e1183cb646a14de18a22539302