Attempt 2 at landing this patch. It broke the Mac/linux builds. Fix for now is to add the offending code in np_v8object.cpp in a ifdef OS_WIN
This fixes http://code.google.com/p/chromium/issues/detail?id=2472, which
is an issue with popups displayed by the flash plugin in response to a user
click, getting blocked.

The plugin invokes NPN_Evaluate to execute the javascript. It also invokes
the NPN_PushPopupEnabledState/NPN_PopupEnabledState API's to set the popup
enabled stack for the duration of the call. 

The fix is to add the plumbing in NPN_Evaluate to pass in a flag indicating
whether popups are allowed for the duration of the call.

Bug=2472
R=jam

Review URL: http://codereview.chromium.org/6379

git-svn-id: http://src.chromium.org/svn/trunk/src@3119 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e527ea843e095779c26c7fd8fc15f1f42bf15c8f