Re-add code to load the ahem font.
We used to load the font.  Then came the font hacks from afm files
which made it no longer necessary to load the font.  The code to
load the font was removed.  Then we got rid of the font hacks and
rebaselined everything.

There are lots of updated results (hundreds), but I'm not uploading
them to rietveld.  I used grep to figure out which tests have
"ahem" in them and rebaselined them.

Review URL: http://codereview.chromium.org/7097

git-svn-id: http://src.chromium.org/svn/trunk/src@3304 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 52f5024b858328f54a74c07146562b67799949cc