Make plugin processes (except Gears) run at normal priority.
This fixes a lot of the spurious unresponsive plugin dialogs, and also makes it so that the browser process isn't hung waiting for a low priority plugin process when cpu usage is high.

I was able to simulate this by using 100% cpu in a test program.  Before this change, the dialog always comes up and a flash video never plays.  The browser process is also hung.  With this change the video plays and I can interact with the browser.
Review URL: http://codereview.chromium.org/7129

git-svn-id: http://src.chromium.org/svn/trunk/src@3323 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 66c75bde124ff567cb0c7b112f629fcc77b0b971