Lots of ScriptController fixes, but it still doesn't compile cleanly.
This makes the header and C file noticeably more like the upstream versions.  In some cases I have ripped out the implementations of functions that we added that there's no obvious way to make it work.  The V8 team is going to have to implement these.

This code absolutely isn't going to run.
Review URL: http://codereview.chromium.org/7313

git-svn-id: http://src.chromium.org/svn/trunk/src@3356 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e2bd7062500652471650b426b0f6377f582000b7