When a mixed-content resource was loaded, we would apply the mixed-content state to the active entry, which would be the pending entry if there is one. That caused a problem in cases where a bad resource is loading and we are initiating a navigation from the browser (and therefore creating a pending entry).
More details:
In the GoodFrame test, the top-frame is HTTPS and we load an HTTP frame with an image. As soon as the bad frame has loaded, we navigate back (creating a pending entry). In rare cases, the notification for the mixed-content image of the HTTP frame would happen after the back action, mistakingly setting the mixed-content state on the pending entry.

BUG=2742
TEST=Run the SSL UI tests 1000 times.
Review URL: http://codereview.chromium.org/7440

git-svn-id: http://src.chromium.org/svn/trunk/src@3480 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 15bcb284fa993c6bc100215edaeb932b307d3a88