Currently, there is no way to access or modify the upload data and/or complete list of request headers associated with a WebRequest object.  The intent is to access this information from WebViewDelegate::DispositionForNavigationAction().
Remove the HasFormData() method that checked for upload (form) data on the
history item. The HasFormData() method is not currently being called from
anywhere, so changes elsewhere are not required.

Add HasUploadData(), GetUploadData() and SetUploadData() methods to test, get
and set the underlying request upload data as a net::UploadData object. Add
GetHttpHeaders() and SetHttpHeaders() methods to get and set all request headers
using a map of values.  Add the SetHttpHeaderValue() method to set the value for
a particular request header.

Patch by Marshall Greenblatt

R=darin



git-svn-id: svn://svn.chromium.org/chrome/trunk/src@3497 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4866bdaf811c2d20bd6caa11f3aa6f366f26fa5c