Construct a field trial to see if HIGH or MEDIUM memory model "works better"
Includes definition of a FieldTrial class to support this.
I have thoughts in my head about how this will eventually
extend to be controllable via UMA (as well as being able
to run tests defined at compile time, as was done in this
example.


r=mbelshe, mmentovai
Review URL: http://codereview.chromium.org/7638

git-svn-id: http://src.chromium.org/svn/trunk/src@3604 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5182bfda6ca2e2fd2acc14cb81902e97cc18f522