Create a thread-safe observer list.  Will be used by SystemMonitor.
Right now the class requires that Observers be RefCounted<>.  This is because we invoke tasks via NewRunnableMethod for them.  However, because we manually track lifecycle via AddObserver/RemoveObserver, we could override the RunnableMethodTraits to not require RefCounted<>.   This would have the advantage that callers do not need to make all Observers be RefCounted, but makes it more critical that observers not forget to call RemoveObserver().
Review URL: http://codereview.chromium.org/7353

git-svn-id: http://src.chromium.org/svn/trunk/src@3787 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2d0e66069084b57b60397ef659617bcc18adba64