Fix Issue 1356872 "Russian bdic is corrupted". The AffReader::AddAffix() function does not reencode a "stripping prefix (or suffix)" field of a PFX (or SFX) line. Unfortunately, this function creates corrupted prefix (and suffix) rules for some dictionaries whose encoding is not UTF-8 (e.g. Russian, Polish, etc.) because a "stripping prefix (or suffix)" field isn't only a length of a prefix (or suffix) but also it is a string to be replaced.
To solve this problem, this change branch checks the third token of an PFX (and
SFX) line represents a "stripping characters" field or a "cross product" field,
and reencode them only if the third token represents a suffix or a prefix.
Needless to say, we also have to re-create bdics and push them to our download
server to solve this issue. :)  

BUG=1428
Review URL: http://codereview.chromium.org/3183
Review URL: http://codereview.chromium.org/3183

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@3903 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 64a7e861f9f0a6acee6fadb09e276013587d0010