Fix Issue 1356872 "Russian bdic is corrupted". The AffReader::AddAffix() function does not reencode a "stripping prefix (or suffix)" field of a PFX (or SFX) line. Unfortunately, this function creates corrupted prefix (and suffix) rules for some dictionaries whose encoding is not UTF-8 (e.g. Russian, Polish, etc.) because a "stripping prefix (or suffix)" field isn't only a length of a prefix (or suffix) but also it is a string to be replaced.
To solve this problem, this change branch checks the third token of an PFX (and
SFX) line represents a "stripping characters" field or a "cross product" field,
and reencode them only if the third token represents a suffix or a prefix.
Needless to say, we also have to re-create bdics and push them to our download
server to solve this issue. :)  

BUG=1428
Review URL: http://codereview.chromium.org/3183
Review URL: http://codereview.chromium.org/3183

git-svn-id: http://src.chromium.org/svn/trunk/src@3903 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 03a41952625f0e6b2ec4b85f432b79eaf18fa350