Moved the SelfDeletePluginInvokeInSynchronousMouseMove to interactive ui tests as we need to simulate mousemoves here. ui tests run on machines which are locked, causing this test to fail.
Moved the NPAPITester and NPAPIVisiblePluginTester classes to a common file
npapi_test_helper.cc so it can be used from both tests.

Relanding the SetCursor patch. It has already been approved by John.

Description below:-
Proposed fix for http://b/issue?id=1362948, which is a crash in the rendererwhen we invoke the setCursor call on the parent view in WebPluginImpl::handleEvent.
This crash occurs because the plugin is deleted in the context of a mouse down event. This could occur by invoking a javascript function via NPN_Evaluate. On return from the HandleEvent sync call we attempt to retreive the parent frame, which
returns NULL and hence the crash.

The fix is to retreive the parent frameview at the start of the WebPluginImpl::handleMouseEvent function and use it whereever needed.

Added a unit test which deletes the plugin instance in a mousemove event.
R=jam
Bug=1362948


Review URL: http://codereview.chromium.org/8691

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@4115 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d9ec9334647f24bc959c34f237fffda402c19100