This fixes http://code.google.com/p/chromium/issues/detail?id=3769, which is to allow users to download the java plugin via our plugin installer.
The plugin installer currently supports URLs which point to executable files.
However the java runtime path is timed and changes frequently. As a stop gap
fix for this, I added a facility in the plugin installer which indicates if
a URL is a display URL. If yes then we navigate to this URL in a new tab,
and allow the user to download the plugin from there.

The format of the chrome_plugins_file.xml has been changed to have each plugin
node specify a list of mime types. This reduces a great deal of duplicated 
information in the plugins file.

Changed the plugin finder URL to http://dl.google.com/chrome/plugins/plugins2.xml
to ensure that older clients continue to work.

Bug=3769
R=jam

Review URL: http://codereview.chromium.org/8729

git-svn-id: http://src.chromium.org/svn/trunk/src@4260 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5d582d459bb36179fd75d9c40a9845505c2ee576