Make ImageBuffer creation and handling conform to proper WebKit form rather than using tacked-on functions on GraphicsContext. Clean up the associated GraphicsContext objects.
Merge SkGraphicsContext, SkPaintContext, and PlatformContextSkia into one class, PlatformContextSkia. This tries to change no code related to this, the code in PlatformContextSkia is just a concatenation of those existing files, and hopefully doesn't need much review. This file is in the wrong style and is a terrible mess, I will clean it up in successive passes so it can be reviewed reasonably.

Review URL: http://codereview.chromium.org/8754

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@4270 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 606f4e2d3af5bad9f3f2d279cbe9b7c08b985b7e