2nd part of the reorganize options under 'UnderTheHood' (bug  3655):
The link 'learn more' now points to a yet-to-be-created article in the help center. I moved it also, so that it appears to the right of the label (instead of below the label).

And finally, I added missing definitions for the Bookmark Manager, which I noticed while I was changing this.

Review URL: http://codereview.chromium.org/8777

git-svn-id: http://src.chromium.org/svn/trunk/src@4317 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fc2eb501267192e28fbf1de97168446a4639eb58