Eliminate CursorChromium's dependency on webkit/glue.
Also modified WidgetChromium to use ChromiumBridge instead of talking to ChromeClientChromium.  I want to eliminate that fake interface in favor of just having our code talk directly to ChromeClientImpl, but that means a dependency on webkit/glue, so I needed to use ChromiumBridge.  Long-term, I'd like to propose changes upstream to HostWindow and ChromeClient to avoid this usage of ChromiumBridge.

The most impactful part of this CL is the change to move the enumeration of cursor types from WebCursor to PlatformCursor.  This means that WebCursor consumers no longer have access to the type enumeration.  I replaced that with helper functions on WebCursor.  I think the result not only achieves the goal of breaking CursorChromium's dependency on webkit/glue but is also much cleaner.

R=iyengar,eseidel
Review URL: http://codereview.chromium.org/9072

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@4846 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5a8a9cf372afe9b8fbad87ebb1b000461d3d2415