This fixes bug http://code.google.com/p/chromium/issues/detail?id=4137, which was an issue with the site failing to detect Windows media player. The site would run a script which would enumerate the list of plugins from the navigator object and compare each plugin name for "Windows Media" and "Plug". This is under the assumption that the windows media player plugin is typically named as Microsoftï¿½ Windows Media Player Firefox Plugin and it would find both the strings in the name.
I had changed the name of activex shim instance for media player previously to Windows Media to fix a similar issue with bloomberg.com. Should have changed it to the full name at that point.

The fix is to fully match the windows media player plugin name in the shim.

The other change is to perform a case insensitive comparison of the plugin file names while looking for matching plugins.

R=jam
Bug=4137

Review URL: http://codereview.chromium.org/9482

git-svn-id: http://src.chromium.org/svn/trunk/src@4950 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d09da99bcc4624b796b6d8621ce867ac1a5b4522