Make the buildbots green.
A test list was accidentally commited with a line commented out.

Aditionally, some of the animation tests are flaky.  When running all
tests, one of the tests seems to consistently time out.  When running
only the animation tests another test is timing out.  Updating the
test list with the test that currently times out on the buildbot.

Review URL: http://codereview.chromium.org/10211

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@5009 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c4effd18a1e856c98150fb74d40c7b89a2623f63