Return the right error code when we proxy a call to the broker.
IIRC we decided to always return access denied because
we did not want to leak the real error code, but this
is bogus for 2 reasons:
1. The broker will return access denied if it's not
allowed in the policy
2. The check to hide the return code is in the renderer,
so it would have been possible for a malicious user to see
it anyway.

I also added a test for it.

BUG:3965

Review URL: http://codereview.chromium.org/10615

git-svn-id: http://src.chromium.org/svn/trunk/src@5388 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0946e0b90df762e7c0015be6975dbfb3148bf9ef