Fix a crash that occurs during the browser shutdown process.
It is possible that during the shutdown process, the database thread
posts tasks back to the IO thread which result in calls to the
SafeBrowsing protocol manager object. This object will have been
deleted by the time the posted task runs, resulting in a crash.

This change checks the enabled_ flag (which is set to false when
the shutdown starts) for all tasks posted to the IO thread to
protect against access to deleted objects.

BUG = http://crbug.com/4531

Review URL: http://codereview.chromium.org/11245

git-svn-id: http://src.chromium.org/svn/trunk/src@5644 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: efefc790c49b9bd73ef0adc68c900837bec03a7c