Fix the getintidentifier-special-values.html plugin layout test, which broke as a result of change r5829 to fix silverlight windowless plugin issues.
One of the changes was to return int32 as the variant type wherever applicable. Firefox does the same and Silverlight does rely on this behavior. This is broken in webkit. I will be filing a bug against webkit.

The fix is to support int32 as an incoming NPVariant type in the layout test plugin.

TBR=jam

Review URL: http://codereview.chromium.org/11574

git-svn-id: http://src.chromium.org/svn/trunk/src@5837 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6f6de55e47bcce7fd48337635ad6c0539e5ca311