faster scons builds of test shell By reducing the number of include paths that are specified as $WEBKIT_DIR/port, we reduce the number of paths that get expanded by the addRepository functions.  Instead, we only specify directories from $WEBKIT_DIR/port if we have header files in those directories.  The down side is that if we add a header to port, we need to update the include paths.
Also, don't specify WebCore includes with OBJ_ROOT as the fallback repository so we save on -I expansion.

Remove $WEBKIT_PORT_DIR because it's the same as $PORT_DIR.

The change to glue/SConscript is to pick up webkit_version.h since glue is the only place that needs it.

Review URL: http://codereview.chromium.org/11598


git-svn-id: http://src.chromium.org/svn/trunk/src@5923 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 086d24a15469f0e976dca116ef7569c447a8827a