Ensure that the context menu shows up in a windowless Silverlight plugin instance.
This fixes bug http://code.google.com/p/chromium/issues/detail?id=4691

The windowless instance of the Silverlight plugin calls the WindowFromPoint API and passes the window handle returned as the owner window in the TrackPopupMenu API call. The TrackPopupMenu API fails if the owner window does not live on the same thread as the caller. It works in the other browsers as the window where the windowless plugin is embedded is in the same thread.

The fix is to add a quirk specific for the TrackPopupMenu issue, IAT patch the Silverlight plugin and pass in the handle to the dummy activation window as the owner.

The other fix made is to have a focus window for the duration of the WM_RBUTTONDOWN/WM_RBUTTONUP sequence. This ensures that keyboard navigation works in the context menu displayed by the Silverlight plugin.

Bug=4691
R=jam

Review URL: http://codereview.chromium.org/12626

git-svn-id: http://src.chromium.org/svn/trunk/src@6029 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 400429e2377a698baff82687834ce692bc204c3e