Get back the messages in the inspector console.
The unforking of the inspector JS files caused the message text not showing in the inspector console due to changes in the API between the inspector controller and the inspector. Manually added the missing changes from WebKit changeset 36817 (http://trac.webkit.org/changeset/36817) to InspectorController.cpp and fixed the parameters to the WebInspector.ConsoleMessage constructor.
Review URL: http://codereview.chromium.org/10997

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6110 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fc4a859597b0e47dbb443ca9d6ce019193aca619