Fix the painting problem with Stree View and Flash 10.  After the size of the plugin changes, we generate a new backing store.  Flash 9 invalidates the rectangle, but Flash 10 doesn't.  As a result we paint from a blank backing store.
The fix is to keep track of which part of the backing store has been painted into by the plugin, and if the renderer tries to paint an area that falls outside of it, do a synchronous paint.

Also get rid of first_paint_, as it was introduced during a failed attempt to make plugin loading asynchronous.
Review URL: http://codereview.chromium.org/10792

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6128 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: cca67d126603bd36d87a2b6d99414a270ba54b8f