Linux: When asked for Monaco font, default to Times New Roman.
Monaco is an Apple specific font which we don't have on either Linux nor
Windows. When Windows is asked for it it appears to fallback to Times
New Roman. We configure Linux to do the same so that we match on
layout-tests which use Monaco.

Review URL: http://codereview.chromium.org/11610


git-svn-id: http://src.chromium.org/svn/trunk/src@6146 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f44e15bb67ff41467ac654aacab9dd715f17b63a