Fix issue 4898: File extension wrong for saved files
When the user specifies a filename without an extension, such as "foo" instead of "foo.gif", we would append "gif" to the filename instead of ".gif" (so the user ends up with a file named "foogif").

I'm not confident this is the only thing wrong with this function. For example if the user specifies "foo." and the filter says "*.gif" shouldn't we add ".gif" to the filename instead of deleting the dot?

Also, we seem to be ignoring the default_extension parameter passed in when we figure out what extension to tack on and instead get the file extension from GetFileExtensionFromPath... :s

Review URL: http://codereview.chromium.org/12838

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6174 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: dc3fb59059806efd6074e81f7b5bc400c2fc17f0