Local fix to SCons bug:
If a Node exists explicitly in a repository directory but not in the
local working directory, SCons would throw an exception when globbing
the entry when it tried to extract the non-existent same-named entry
from the local working directory's entries dictionary.  Since we know
we want to return local entry anyway, make sure it exists.
Review URL: http://codereview.chromium.org/13044

git-svn-id: http://src.chromium.org/svn/trunk/src@6212 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1cae06d10c64382ae0c8fdf4e889f1848fcab585