Allow PDF files to open in Acrobat Reader. This fixes http://code.google.com/p/chromium/issues/detail?id=3977, which is an issue with Acrobat Reader internet preferences being set to always open PDF files in Reader. This operation when performed in chrome and Safari brings up an error message from the Acrobat Reader plugin that the file could not be found.
The plugin in this case returns NP_ASFILE as the plugin mode from NPP_NewStream. The data is written accordingly to the temp file created. When we finish receiving the data we pass the file name in NPP_StreamAsFile to the plugin. The plugin does not open the file immediately as per our expectation. It opens the file after the stream has been destroyed which ends up deleting the temp file created above.

Debugged Firefox and found that it never deletes these temp files and leaves them around in the cache. 

We can go one step forward and track all these files on a per instance basis. These are deleted when the plugin instance is destroyed.

Bug=3977
R=jam

Review URL: http://codereview.chromium.org/12875

git-svn-id: http://src.chromium.org/svn/trunk/src@6245 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 377788707e489728494ee2edc6933728cc9546d9