Linux: calculate fractional xheight
We need to push the calculation of the x-height into Skia. For one of
the layout tests, it assumes that 2.5ex of Ahem is exactly 32px.
Previously we calculated the x-height of Ahem to be 13px, so 2.5*13 was
33px.

Now we use the 26.6 fixed point number from freetype's hinter and turn
that directly into a floating point value (x-height is the only WebKit
metric which is a floating point value).

This lets us pass LayoutTests/css2.1/t1507-c526-font-sz-02-b-a.html
(although the Windows baseline is wrong)

Review URL: http://codereview.chromium.org/13112


git-svn-id: http://src.chromium.org/svn/trunk/src@6344 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: bb16745fdf0e73f760b1ac2cd0ab21052bc1ced5