Correctly handle (and log) receipt of gzip response to SDCH encode proposal
When the only response to a SDCH encoding proposal (including a statement of
an Avail-Dict) is a gzip response, the filtering infrastructure needs to
add in a missing "sdch" sniffing filter.  This filter should enter into
pass-through mode if the content does not have a valid SDCH header.  This
functionality is critical to the latency experiment, which will signal
the fact that a session is in the control group by doing exactly the
above pattern (only using gzip, despite the dictionary).

This is a blocker for the Chrome SDCH latency experiment.


bug = 1520081
reviewers= huanr,openvcdiff,kmixter
Review URL: http://codereview.chromium.org/12964

git-svn-id: http://src.chromium.org/svn/trunk/src@6415 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5d64b476f1a09d85b095f0080911a059dd44cc1c