Fix a race condition in the SyncChannel code.
The problem is that QueueMessage would be called with listener() on the IPC thread, and then the function would wait on message_lock_ before adding the message and its associated listener.  However, in between the time that listener() was called and the lock was taken control of, the listener thread may be calling RemoveListener.

The fix is to not send the listener, but instead just send a pointer to the context.  It'll be used on the listener thread, at which point listener() can be called to check if the listener is available.

BUG=1305036
Review URL: http://codereview.chromium.org/12952

git-svn-id: http://src.chromium.org/svn/trunk/src@6448 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9147d3c9fea7491cfb72b797fb8f78a4aed984dd