Linux: checkin rebaselined images for LayoutTests/fast/*
This is the set of images and checksums for layout tests where we match
the render tree exactly and the fuzzy image matcher believes that our
image outputs are semantically the same as the Windows version.

Since we don't want to be running the fuzzy matcher forever, we are
checking in new baselines for the images only.


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6471 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 509764387910be5bf328dc017bf20feb4bda8afe