Linux: checkin rebaselined images for LayoutTests/fast/*
This is the set of images and checksums for layout tests where we match
the render tree exactly and the fuzzy image matcher believes that our
image outputs are semantically the same as the Windows version.

Since we don't want to be running the fuzzy matcher forever, we are
checking in new baselines for the images only.


git-svn-id: http://src.chromium.org/svn/trunk/src@6471 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 31e20ae818e978946e9832a5986f1202e353144b