Linux rebaseline: LayoutTests/css2.1/*
This adds new image baselines for tests under css2.1/* where either the
fuzzy image matcher accepts the delta, or where the images have been
manually checked.

This also adds render-tree baselines for 8 tests which current fail
because of differences in the arrow or ying-yang glyph between Windows
and Linux. On Linux we are using Lucida Sans for the symbols and it's
not clear what Windows is using, but it's not a free font.


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6631 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9e7fa703f7f3f982ecbf461b104a1c3a46e13e64