Disk cache: Set up an experiment to measure the effect of increasing the maximum cache size.
There will be three groups on the experiment, each group with
10% of the users (Dev Channel). Each group will have the max
cache size increased by a factor from 2 to 4 times the current
size, and we'll measure the age of evicted entries.

I'm also adding a few metrics of reliability, to detect how
often we're not getting to do proper cleanup.

Review URL: http://codereview.chromium.org/14013

git-svn-id: http://src.chromium.org/svn/trunk/src@6847 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 45c699462e158c887bc85bddcfa3d79d6e2fef24