Disk cache: Set up an experiment to measure the effect of increasing the maximum cache size.
There will be three groups on the experiment, each group with
10% of the users (Dev Channel). Each group will have the max
cache size increased by a factor from 2 to 4 times the current
size, and we'll measure the age of evicted entries.

I'm also adding a few metrics of reliability, to detect how
often we're not getting to do proper cleanup.

Review URL: http://codereview.chromium.org/14013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6847 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5ec86111e652b944016de509b4838bdfcc5e227c