Second attempt at extending the back button to the left edge when maximized.
This time, we do everything programatically and extend the target area only when the window is maximized.  After this patch, we don't actually get the left-most pixel to be targetable on XP because our frame code is borked, but this should work once we unbork that.

BUG=3484
R=beng

Review URL: http://codereview.chromium.org/13782

git-svn-id: http://src.chromium.org/svn/trunk/src@7005 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 57ef5efbab0944a642d0388b5a0360ad9ff03c2e