Disk cache: remove the hard coded list from rankings.cc
The rankings module now works with any list, not just
the first one. This is part of the support needed for
new eviction algorithms.

note: Rankings::CheckList() is basically the code that
was implementing the old Rankings::SelfCheck() (moved
with this cl).

The disk cache behavior is not changing with this cl.

Review URL: http://codereview.chromium.org/14141

git-svn-id: http://src.chromium.org/svn/trunk/src@7074 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 258f7208ca6807dd6e15e0bf85c295e7614501e0