Disk cache: remove the hard coded list from rankings.cc
The rankings module now works with any list, not just
the first one. This is part of the support needed for
new eviction algorithms.

note: Rankings::CheckList() is basically the code that
was implementing the old Rankings::SelfCheck() (moved
with this cl).

The disk cache behavior is not changing with this cl.

Review URL: http://codereview.chromium.org/14141

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7074 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2cc52bcf9e17ba335c9945c9a61c54079a7a574c