Reverting 7296.
From http://build.chromium.org/buildbot/waterfall/waterfall?builder=Chromium%20Reliability
There is a clear increase of crashes after this change. It really seems like a heap overflow but it's hard to tell. Reverting to see if it helps the reliability tests. If not, we can reapply this change and find another culprit.
Review URL: http://codereview.chromium.org/15081

git-svn-id: http://src.chromium.org/svn/trunk/src@7304 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c9a00ca2891a0b3e21a3acd6f36b93bf5436d8b7