Fix 4830: We navigate back after closing the Save As dialog.
If you use Alt-Shift-T and the arrow keys to go to the Page menu 
and then launch the Save Page As dialog, we would navigate back after 
saving the page.

This is because once we close the Save As dialog with Enter, focus
goes to the Back button (which arguable is also wrong, see issue 5750)
and it receives a KeyRelease(Enter), which causes us to navigate back 
right after saving the page.

In Windows, pressing Enter while a button has focus clicks the button
on KeyPressed, whereas we currently click the button on KeyRelease. 

This patch makes Chrome mimic native Windows behavior. We retain
the current behavior for Space, however, which clicks on KeyReleased 
in both Chrome and Windows.

BUG=4830
TEST=Native Chrome buttons should click buttons on KeyPress when using
Enter and on KeyRelease when using Space.

Review URL: http://codereview.chromium.org/16202

git-svn-id: http://src.chromium.org/svn/trunk/src@7406 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fe793f66fa413c0de53aa4442bfe63de3a7c7ed7