Update Channel Changer to support a new Dev channel tag.
I want to declare 1.1-dev users "on the Beta channel" and
create a new 2.0-dev tag for the Dev channel. To do that 
I needed to make the Channel Changer recognize 1.1-dev as
Beta.

I also changed some strings:

  * make the note about changing to a more stable channel 
    a little clearer
  * make the dialog title 'Channel Changer' (instead of 
    Switcher)

The vcproj needed to be updated for the Release rule. 
(Jungshik changed Debug in r7532.)

Review URL: http://codereview.chromium.org/17050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7559 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ea5e4233898902f2a40f2f5b0c729082e5e06c6b