Update Channel Changer to support a new Dev channel tag.
I want to declare 1.1-dev users "on the Beta channel" and
create a new 2.0-dev tag for the Dev channel. To do that 
I needed to make the Channel Changer recognize 1.1-dev as
Beta.

I also changed some strings:

  * make the note about changing to a more stable channel 
    a little clearer
  * make the dialog title 'Channel Changer' (instead of 
    Switcher)

The vcproj needed to be updated for the Release rule. 
(Jungshik changed Debug in r7532.)

Review URL: http://codereview.chromium.org/17050

git-svn-id: http://src.chromium.org/svn/trunk/src@7559 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 00b743b3ff37313f6e2b7c3c6666619b6a62fed7