This CL adds a way to block resource requests in the ResourceDispatcherHost for specific RenderViewHosts. This is used by the interstitial code to prevent the original page from making network requests while the interstitial is showing.
Several UI tests are still required to test this. Because of the inherent complexity of the scenarios to test this, I'm afraid these tests are going to be flakey and soon after disabled (like many of the interstitial UI tests at this point).
This will be done next as part of my next effort to mock some of the WebContents stuff.

TEST=Run the unit tests. Create a page that does XMLHttpRequests and log when the requests complete with the time it did. While that page is shown, switch to a malware/bad SSL page. Come back to the page, ensure no requests were processed while the interstitial was showing.
BUG=None
Review URL: http://codereview.chromium.org/16546

git-svn-id: http://src.chromium.org/svn/trunk/src@7898 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d66eba18b1936f6c0e12cd563b7dcff9841faa85