On exit do not terminate renderers hosted in external tabs
When a browser window is closed, we just terminate
all existing renderers for fast shutdown. If we have
any renderers hosted externally using automation they
are also killed during this process. 

The fix is to prevent fast termination if any of the
tabs are hosted by automation.


Review URL: http://codereview.chromium.org/17453

git-svn-id: http://src.chromium.org/svn/trunk/src@7901 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f2ab113ea1f4c6ef9ce3887757766b7eae47ca05