Pixel perfection on OTR icon.  The tabstrip is still positioned/sized wrong relative to the OTR icon, sadly.
Much of the issue here is because the OTR image was a child view in the old code and isn't in the new.  As a result, we have to:
* Change the drawing order to draw the OTR image after the tabstrip (like a child view would normally have been drawn), so the tabstrip's semi-transparent pixels don't overlay it
* Change the bounds calculation and drawing code to clip the image properly in maximized mode (where a child view would have been clipped naturally)

Since I changed the calling order of the OTR functions, I also changed their declaration/definition order to follow suit.

This also tries to rename a few constants to be more consistent.  I can't figure out how to make any of these read nicely :(

BUG=5054
Review URL: http://codereview.chromium.org/17362

git-svn-id: http://src.chromium.org/svn/trunk/src@7908 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9bbb293cfdf84d4c518d320c0911a1af07dd6e9b