Report name of repeatedly registered histograms
Histograms should ONLY be listed once in the code
for a given process (the browser process?).  If 
two or more histograms have the same name, it is an
error, and a DCHECK will fire.  This code will also
display the name of the offending histogram.

r=sky
Review URL: http://codereview.chromium.org/16609

git-svn-id: http://src.chromium.org/svn/trunk/src@8009 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 07c6df9942e77acf898e672c4ea5844caa92fb55