Render maximized mode better when using nonstandard border widths.
This change is a bit gross.  It calls out to GetSystemMetrics() all over this code, which just shows how many places are sizing themselves from the window edge instead of from the window border edge.  When cleaning this file up to be more sane, I intend to make most of these functions position themselves based on the border edge, which should cut down the number of these calls to just a couple (and perhaps collapse some of the constants together too).

BUG=5054

TEST=In both Classic and Luna, use the Display control panel to set your active/inactive border widths to something nonstandard (try 20, that's real noticeable) and maximize a Chrome window.  Make sure the window controls appear where you expect instead of partly or totally off the screen edges.
Review URL: http://codereview.chromium.org/18255

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@8060 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 39174ce53f5e069ab31111b4a8a0c6614a02775a