Add data pack file format to the grit output type and enable it for webkit_resources.grd.
This only adds support for <include> nodes.  <message> nodes are forthcoming.

The design doc describing the format is here:
http://dev.chromium.org/developers/design-documents/linuxresourcesandlocalizedstrings

I hooked into the <release> tag because we need to generate an index at the beginning of the file.  The index contains the position of all the resources, so we walk the file at that point.

This currently only handles <include> tags, but will be extended to handle <message> as well.

Review URL: http://codereview.chromium.org/18215


git-svn-id: http://src.chromium.org/svn/trunk/src@8147 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c743291ce943bcd8777117b1ebc434edb9c471da