Use PlatformThreadId, not int when dealing with thread ids.
Windows uses a DWORD (unsigned long) for thread ids and POSIX uses a
pid_t (int on Linux) for the same. In the code, we are currently
stuffing thread ids into an int which is dangerous on Windows (because
DWORDS can exceed an int and wrap) and will break if pid_t is ever !=
int.

This change changes all the places where we currently have an int to
use a new typedef, PlatformThreadId. This change also needs to occur
for process ids, but I'm not doing that in this CL.

Review URL: http://codereview.chromium.org/18677


git-svn-id: http://src.chromium.org/svn/trunk/src@8525 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 92dfc3553bd50f5f258a268e9c0cd43f2960c3bc