Handle race between deleting a low use referrer acruing value
Prevent a DCHECK from firing.

As we adaptively learn about referrers to enhance DNS
pre-resolution, we are forced to evict some entries
to prevent accidentally getting really large lists
(when we make mistakes).  Sometimes we evict a 
referred item from a list *while* that item is being
resolved (because it was in teh list).  In that case,
we have to be careful when we try to acrue credit 
(latency savings).

r=paulg
Review URL: http://codereview.chromium.org/18527

git-svn-id: http://src.chromium.org/svn/trunk/src@8531 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 121707bbac2fd387c4186bcffa1d7778a9fbae4c