Separate out the backing store from the RenderWidgetHost into its own file to make porting easier.
I also did some cleanup in the scrolling and painting areas, collapsing the
confusing ScrollRect call inside OnMsgScrollRect to just be inside the function
and call a new function on the backing store to scroll it. Same for painting.

This also moves plugin window moving to the view.
Review URL: http://codereview.chromium.org/18702

git-svn-id: http://src.chromium.org/svn/trunk/src@8594 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4bc567a1f7b1d2642acb218fc6f7dddd1d2c919f