Add user script support to extensions.
This is implemented mostly by relying on the existing
user script code. But since extension user scripts are
declared, not discovered in a directory, I had to add
support for adding 'lone' user scripts to
UserScriptMaster. This led to a bit of refactoring.

Note that this CL relies on:
http://codereview.chromium.org/18352

Review URL: http://codereview.chromium.org/18198

git-svn-id: http://src.chromium.org/svn/trunk/src@8614 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 102634c7b7b294daf4c3294ae7f55c96f3e5218b