Index: /usr/local/google/home/shess/chromium/chromium-svn-submit/src/webkit/tools/layFollowup from http://codereview.chromium.org/18738
Went through looking for Mac-specific CRASH tests, and ran them with
and without the above change.  These were the ones which went from
crashing w/out the change to not-crashing with the change.  Since the
change fixed a race condition, I also replicated the crash for these
ones w/in gdb, and the crash point was consistent with the above fix.


git-svn-id: http://src.chromium.org/svn/trunk/src@9204 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 556d4459cb5164278b0ed3275312171c30315744