Fix Renderer Crash on OS X + update unit test. The problem was that the process shutdown_event_ was created in RenderProcess::GlobalInit() and then deleted in RenderProcess::GlobalCleanup() and was being accessed after it's deletion by the MessageLoop's destructor (which was trying to signal shutdown).
MessageLoop was still retaining a reference to it since there was still an IPCSyncChannel listening on it.

IPCSyncChannel hadn't stopped listening on it since it was being terminated via OnChannelError.

This fix adds code that terminates listening on a channel when a Channel error occurs.

Review URL: http://codereview.chromium.org/20114

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@9295 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 393b4e7c553831092dab80a1ba529684b98e521f