Make the mime -> extension code return a dotless extension.
Previously we would return something like ".html", because that's how it's stored in the registry on Windows.  For a variety of reasons, it's much simpler to work without the prefixed dot.  In fact, the only two consumers were manually stripping of the prefix dot.

There is a cooresponding change that should happen in WebKit / ChromiumBridge, but since it's checking if there is a prefix dot and removing it, the code will still work, and I'll follow up with that change.

Review URL: http://codereview.chromium.org/23010


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@9333 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e442a702651bd8f25ac8f17b1df726c3edb7dd5f