Call the DidStartLoading/DidStopLoading methods on the WebViewDelegate interface whenever we receive the initial response for a byte range request and when the request eventually completes.
This ensures that the throbber continues to spin when these requests take a while. 

It stops spinning during load when the document load is cancelled by the plugin process. This occurs when the plugin invokes NPN_RequestRead the first time on the stream.

This fixes http://code.google.com/p/chromium/issues/detail?id=7375

Bug=7375

Review URL: http://codereview.chromium.org/20333

git-svn-id: http://src.chromium.org/svn/trunk/src@9732 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e87cd8208fd6edd49c344d968dcc3be4513afd20