Call the DidStartLoading/DidStopLoading methods on the WebViewDelegate interface whenever we receive the initial response for a byte range request and when the request eventually completes.
This ensures that the throbber continues to spin when these requests take a while. 

It stops spinning during load when the document load is cancelled by the plugin process. This occurs when the plugin invokes NPN_RequestRead the first time on the stream.

This fixes http://code.google.com/p/chromium/issues/detail?id=7375

Bug=7375

Review URL: http://codereview.chromium.org/20333

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@9732 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 250c29c2488c47eafe97f067397aa22e3644a301