Check for hung plugin windows before detaching them
The process of detaching plugin windows from tab window
involves windows messages across browser and plugin process.
If the pungin process is hung this hangs the browser 
shutdown sequence. 

This workaround first checks if the plugin is responsive
before detaching its window from the tab. If a plugin is
hung, it's window would not be detached (and would be
destroyed with the tab window). 

The side effect of not detaching is that some plugins
expect their window to be valid until they get NPAPI 
shut down notifications and may crash if those 
notifications were sent after the window was destroyed.
But since the plugin is hung anyway, we can go ahead
not detaching the window. Plugin process will be 
cleaned up later.

BUG=3613



Review URL: http://codereview.chromium.org/20301

git-svn-id: http://src.chromium.org/svn/trunk/src@9756 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2b8ea24c5bd14645627e343907ebf95e0ed8357b