Revert my SetWindow(NULL) change in r9653.
I am told we don't want to change any NPAPI behavior on Windows, and that we should try to match Safari's behavior.

This change is neccessary for plugins on Linux, so the code will have to be forked and this patch reapplied.

This is a little more than a plain revert, since r9692 moved around some code.

Review URL: http://codereview.chromium.org/21405


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@9873 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 209ed0e7dc5d645a5186a57db75005e1df35ffc4