Include a better sample JSON dataset in traceline.
The previous startup-off2.json didn't have stack traces.  It was also quite large, it was unfiltered and contained every single system call event.

Include a program to filter a JSON data set to only include events that took a worthwhile amount of time.  This greatly reduces the size of the JSON data.

A new filtered JSON dataset startup-release.json is included, much smaller and with stack traces.

Added an ignore for checkdeps on traceline/.

Review URL: http://codereview.chromium.org/20535

git-svn-id: http://src.chromium.org/svn/trunk/src@10094 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1dd94e6a283d2f142691624340afc34fcf192333