ï»¿This CL fixes issue 2674 - Omnibar displays wrong parenthesis for "arabia.msn.comâ€¬" (http://crbug.com/2674)
The issue is: in RTL locale, the search engine name, if ending with punctuation, might not be displayed correctly in omnibox keyword text and omnibox keyword hint.

Fix: adjust the search engine keyword description (short name) for RTL locale so that the pure LTR search engine name 
is marked with LRE-PDF pair to be displayed correctly. For example, "MSN(English)" is displayed as is, not as "(MSN(English".

Reproduce steps:
Looks like the "MSN(English)" is no longer a default search engine name. We can make one by the following steps:
1. Open "customize and control Google Chrome" --> "Options",
2. under "basics" tab, under "Default search:", click "Manage"
3. click on one search engine from the "default search options", then click "edit". For example, pick "AOL". 
4. change the "name" from "AOL" to "AOL(English)".

Without the fix, in RTL Chrome, when type in "aol.com" in omnibox, the omni keyword hint shows "(AOL(English - xxxx <tab> yyy" while "xxx" stands for "to search" in Hebrew and "yyy" stands for "press" in Hebrew. After press <tab>, the keyword text in omnibox is ":(AOL(English - zzz" while "zzz" stands for "search" in Hebrew.

With the fix, the name "AOL(English)" is displayed correctly in the above 2 places.


Review URL: http://codereview.chromium.org/21320

git-svn-id: http://src.chromium.org/svn/trunk/src@10116 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4cc75b4ea28f038211e5764b54959400c058d2fb