This CL fix the following 2 bugs: 1.7324 -- RTL: Path contains Hebrew is wrong on "Download location" text box (http://crbug.com/7324)
2. 7326 -- RTL: Folder icon missing on "Download location" text box
(http://crbug.com/7326)

The fix are:
1. force path to have LTR directionality.
2. draw icon in mirrored position in RTL locales.

Steps for test:
1. Create a folder with a Hebrew file name on Hebrew XP
2. Change the download location to the folder created in step 1
3. Observe the path displayed on "Google Chrome Options" --> "Minor Tweaks" --> "Download location" text box

Without the fix: 
The path contains Hebrew folder name is wrong. It displayed as "cCBA\:" while the path is "c:\CBA" where "CBA" is a Hebrew folder name.
And there is no folder icon in the "download location" text box.

With the fix:
The path displayed correctly as "c:\CBA" or "c:\CBA\FED" where "FED" is subfolder of "CBA".
And the folder icon showed at the very right.


Review URL: http://codereview.chromium.org/20038

git-svn-id: http://src.chromium.org/svn/trunk/src@10121 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2232ffec0e0a090f44c7fb1ddd027da1cc57d193