Re-re-roll use of ongoing logs to gather stability metrics
This was rolled back to try to use only the initial logs to gather statibility
stats, because we were having trouble capturing data in ongoing logs.

We now resolved the problem (which was actually only on the stability
branch), so we'll roll back to using this method and verify it in the
dev branch.

tbr=mal
Review URL: http://codereview.chromium.org/27061

git-svn-id: http://src.chromium.org/svn/trunk/src@10238 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8cfe7ac0f3753b975c3f6111f124eca479cf82e8