Fix grit include dependencies in the scons build.
Re-enable the scons code that walks a grd file and adds its dependencies as
build dependencies.  This requires that all our dependencies are files in our
source tree (which is what motivated getting rid of the tld data as a
generated resource).  I had to fix up the terms of service for this.

Also remove some code that disabled .pak file generation on windows which was
broken because it couldn't find the tld data.

Review URL: http://codereview.chromium.org/28194


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10494 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6f7af59547a46c821ac6a04b01c6c1689787937a