Enable --single-process for Mac and Linux.
Additionally, wire in a bit of setup for the --single-process case,
because it cannot be done by the renderer thread because that thread
is not the main thread.

Fair warning: --single-process often seems to be broken by various
unrelated changes, which may-or-may-not make assumptions about the
process architecture.  I will try to stay on top of these.

Review URL: http://codereview.chromium.org/21330

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10559 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fdd6077dd5f390ac08e7e13fa25447847c5c5730