Enable --single-process for Mac and Linux.
Additionally, wire in a bit of setup for the --single-process case,
because it cannot be done by the renderer thread because that thread
is not the main thread.

Fair warning: --single-process often seems to be broken by various
unrelated changes, which may-or-may-not make assumptions about the
process architecture.  I will try to stay on top of these.

Review URL: http://codereview.chromium.org/21330

git-svn-id: http://src.chromium.org/svn/trunk/src@10559 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4c145a57e3277eb6b933ca52d278049eea146565