Attempt 2 at landing this.
Don't call SetPriorityClass if CBSText.dll is loaded in the browser process. The
CBSText.dll loads as a global GetMessage hook into the browser process and
intercepts/unintercepts the kernel32 API SetPriorityClass in a background
thread. If the browser calls this API just
when it is intercepted the dlls interceptor proc messes up the stack while
returning back, which causes random crashes in the browser.

The hacky fix for now is to check for this dll and not invoke the API. We will
beautify this fix if it works in the wild.

This fixes http://code.google.com/p/chromium/issues/detail?id=6418

Bug=6418

Review URL: http://codereview.chromium.org/27298

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10676 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 66b20f837c20bb26215d1cf0d97fd3854597d502