Fix startup_test failures on Windows buildbot.
The problem seems to be an error when overwriting file which is in use
(that's how EvictFileFromSystemCache works on Windows). So I used
a standard workaround for such Windowsic problems and created a wrapper
which retries 10 times. Similar code was there before my porting patch.

TBR=agl

Review URL: http://codereview.chromium.org/40160

git-svn-id: http://src.chromium.org/svn/trunk/src@10981 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3e078529d3800f450044239726d0989904bf7da3